+++ src/views/Authentication/SignInAuth.tsx
@@ -1,25 +1,25 @@
 import React from "react";
 import { Link, useNavigate } from "react-router-dom";
-import { withAsync } from "../../helpers/withAsync";
-import { signinAuth } from "../../api/AuthenticationApi";
+import { withAsync } from "helpers/withAsync";
+import { signinAuth } from "api/AuthenticationApi";
 import { useDispatch } from "react-redux";
 import { AppDispatch } from "store/store";
 import { toast } from "react-toastify";
-import TeratanyLogo from "../../assets/Teratany_ico/apple-touch-icon-180x180.png";
-import FormField from "../../components/common/FormField";
-import Button from "../../components/common/Button";
+import TeratanyLogo from "assets/Teratany_ico/apple-touch-icon-180x180.png";
+import FormField from "components/common/FormField";
+import Button from "components/common/Button";
 import { Formik, Form } from "formik";
 import * as Yup from "yup";
 import ErrorMessageForm from "components/common/ErrorMessageForm";
 import {
   UserInitialState,
   setAuthentication,
-} from "../../store/reducer/user.reducer";
+} from "store/reducer/user.reducer";
 import jwtDecode from "jwt-decode";
 import { AxiosError } from "axios";
-import useLoadingButton from "../../hooks/useLoadingButton";
-import { getById } from "../../api/ProfileApi";
-import { setAccountConnected } from "../../store/reducer/account.reducer";
+import useLoadingButton from "hooks/useLoadingButton";
+import { getById } from "api/ProfileApi";
+import { setAccountConnected } from "store/reducer/account.reducer";
 
 interface signinFormValues {
   email: string;
@@ -39,7 +39,6 @@ const SignInAuth: React.FC = () => {
   const fetchProfile = async (token: string, id: string) => {
     const { error, response } = await withAsync(() => getById(token, id, id));
 
-
     if (error instanceof AxiosError) {
       const error_message: string =
         error?.response?.data.description ||
@@ -57,8 +56,6 @@ const SignInAuth: React.FC = () => {
       signinAuth(values.email, values.password)
     );
 
-    console.log("error ===> ", error);
-    console.log("response ===> ", response);
     if (error instanceof AxiosError) {
       endLoading();
       const error_message: string =
